pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "instadon"
version = "1.0.0"
description = "Cross-post from Instagram to Mastodon"
authors = [{name = "Your Name", email = "your.email@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "instaloader>=4.14.1",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
]

[project.scripts]
instadon = "instadon.main:main"

[project.urls]
Homepage = "https://github.com/yourusername/instadon"
Repository = "https://github.com/yourusername/instadon"
